requirements.txt
asn1crypto==0.24.0
boto3==1.7.31
botocore==1.10.31
cffi==1.11.5
colorlog==3.1.4
cryptography==2.2.2
docutils==0.14
edlib==1.2.3.post1
idna==2.6
jmespath==0.9.3
pycparser==2.18
PyStaticConfiguration==0.10.3
python-dateutil==2.7.3
PyYAML==3.12
s3transfer==0.1.13
six==1.11.0
